backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
boto3==1.35.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
aiosqlite==0.20.0
pydantic==2.8.2
